requirements.txt
asgiref==3.5.2
Django==4.0.1
django-environ==0.9.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
sqlparse==0.4.3
tzdata==2022.7
djangorestframework==3.14.0
social-auth-app-django==5.0.0
django-cors-headers==3.13.0


